feat(FileManager): refactor to use custom hook for file management logic
