+++ apps/chat/src/components/FileManager/hooks/useFileManager.tsx
@@ -1,0 +1,580 @@
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+
+import { useFileManagerActionLabels } from '@/src/hooks/useFileManagerActionLabels';
+import { useTranslation } from '@/src/hooks/useTranslation';
+
+import {
+  buildFileTree,
+  convertToUIKitFile,
+  filterFilesByFilters,
+  filterFoldersByFilters,
+} from '@/src/utils/app/file-manager-adapter';
+import { getFileRootId, isRootId } from '@/src/utils/app/id';
+import {
+  PublishedWithMeFilter,
+  SharedWithMeFilters,
+  defaultMyItemsFilters,
+} from '@/src/utils/app/search';
+
+import { Translation } from '@/src/types/translation';
+
+import { ShareActions } from '@/src/store/actions';
+import { FilesActions } from '@/src/store/files/files.reducers';
+import { FilesSelectors } from '@/src/store/files/files.selectors';
+import { useAppDispatch, useAppSelector } from '@/src/store/hooks';
+import { SettingsSelectors } from '@/src/store/settings/settings.selectors';
+
+import { FeatureType, UploadStatus } from '@epam/ai-dial-shared';
+import {
+  ButtonVariant,
+  DialCopiedItem,
+  DialDeletedItem,
+  DialFile,
+  DialUploadFileItem,
+  FileManagerColumnKey,
+  useDialFileManagerTabs,
+} from '@epam/ai-dial-ui-kit';
+
+export const useFileManager = () => {
+  const dispatch = useAppDispatch();
+  const { t } = useTranslation(Translation.SideBar);
+
+  const isFileMetadataLoading = useAppSelector(
+    FilesSelectors.selectLoadingFileMetadata,
+  );
+  const areFilesLoading = useAppSelector(FilesSelectors.selectAreFilesLoading);
+  const areFoldersLoading = useAppSelector(
+    FilesSelectors.selectAreFoldersLoading,
+  );
+  const initialDataStatus = useAppSelector(
+    SettingsSelectors.selectInitialDataStatus,
+  );
+  const isAnyOperationInProgress = useAppSelector(
+    FilesSelectors.selectIsAnyFileOperationInProgress,
+  );
+
+  const isCopyingFiles = useAppSelector(FilesSelectors.selectIsCopyingFiles);
+  const isMovingFiles = useAppSelector(FilesSelectors.selectIsMovingFiles);
+  const movingFilesCountRef = useRef<number>(0);
+
+  const fileMetadata = useAppSelector(FilesSelectors.selectFileMetadata);
+  const files = useAppSelector(FilesSelectors.selectFiles);
+  const folders = useAppSelector(FilesSelectors.selectFolders);
+
+  const isLoadingSearchListing = useAppSelector(
+    FilesSelectors.selectIsLoadingSearchListing,
+  );
+
+  const [currentPath, setCurrentPath] = useState<string | undefined>();
+  const [destinationPath, setDestinationPath] = useState<string | undefined>();
+
+  const searchResults = useAppSelector(
+    useCallback(
+      (state) =>
+        currentPath
+          ? FilesSelectors.selectSearchResultsForFolder(state, currentPath)
+          : [],
+      [currentPath],
+    ),
+  );
+
+  const searchResultsUIKit = useMemo(
+    () => searchResults.map(convertToUIKitFile),
+    [searchResults],
+  );
+
+  const [treeCollapsedState, setTreeCollapsedState] = useState<
+    boolean | undefined
+  >(true);
+
+  const { activeTab, handleTabChange, tabs } = useDialFileManagerTabs({
+    my_files: t('My Files'),
+    shared: t('Shared with Me'),
+    organization: t('Organization'),
+  });
+  const previousActiveTabRef = useRef(activeTab);
+
+  useEffect(() => {
+    if (initialDataStatus === UploadStatus.LOADED) {
+      dispatch(FilesActions.getFilesWithFolders({}));
+    }
+  }, [initialDataStatus, dispatch]);
+
+  useEffect(() => {
+    if (currentPath && !isRootId(currentPath)) {
+      const folder = folders.find((folder) => folder.id === currentPath);
+      if (folder?.status !== UploadStatus.LOADED) {
+        dispatch(FilesActions.getFilesWithFolders({ id: currentPath }));
+      }
+    }
+  }, [dispatch, currentPath, folders]);
+
+  useEffect(() => {
+    if (
+      destinationPath &&
+      destinationPath !== currentPath &&
+      !isRootId(destinationPath)
+    ) {
+      const folder = folders.find((folder) => folder.id === destinationPath);
+      if (folder?.status !== UploadStatus.LOADED) {
+        dispatch(FilesActions.getFilesWithFolders({ id: destinationPath }));
+      }
+    }
+  }, [dispatch, currentPath, destinationPath, folders]);
+
+  const {
+    fileTreeItems,
+    rootFolder,
+    loadedFoldersPaths,
+    sharedByMePaths,
+    visibleColumns,
+  } = useMemo(() => {
+    let filteredFiles = files;
+    let filteredFolders = folders;
+    let breadcrumbLabel = t('My Files');
+    const visibleColumns: FileManagerColumnKey[] = [
+      FileManagerColumnKey.Name,
+      FileManagerColumnKey.UpdatedAt,
+      FileManagerColumnKey.Size,
+      FileManagerColumnKey.Actions,
+    ];
+
+    switch (activeTab) {
+      case 'my_files':
+        filteredFiles = filterFilesByFilters(files, defaultMyItemsFilters);
+        filteredFolders = filterFoldersByFilters(
+          folders,
+          defaultMyItemsFilters,
+        );
+        breadcrumbLabel = t('My Files');
+        break;
+      case 'shared':
+        filteredFiles = filterFilesByFilters(files, SharedWithMeFilters);
+        filteredFolders = filterFoldersByFilters(folders, SharedWithMeFilters);
+        breadcrumbLabel = t('Shared with Me');
+        visibleColumns.push(FileManagerColumnKey.Author);
+        break;
+      case 'organization':
+        filteredFiles = filterFilesByFilters(files, PublishedWithMeFilter);
+        filteredFolders = filterFoldersByFilters(
+          folders,
+          PublishedWithMeFilter,
+        );
+        breadcrumbLabel = t('Organization');
+        break;
+      default:
+        break;
+    }
+
+    const { rootFolder, items, loadedFoldersPaths, sharedByMePaths } =
+      buildFileTree(filteredFiles, filteredFolders, breadcrumbLabel);
+
+    if (activeTab !== previousActiveTabRef.current) {
+      setCurrentPath(rootFolder.id);
+      previousActiveTabRef.current = activeTab;
+    }
+
+    return {
+      rootFolder,
+      fileTreeItems: items,
+      loadedFoldersPaths,
+      sharedByMePaths,
+      visibleColumns,
+    };
+  }, [t, files, folders, activeTab, previousActiveTabRef]);
+
+  const getParentPaths = useCallback((path: string | undefined): string[] => {
+    if (!path) return [];
+    const parts = path.split('/');
+    const paths: string[] = [];
+    for (let i = 1; i <= parts.length; i++) {
+      paths.push(parts.slice(0, i).join('/'));
+    }
+    return paths;
+  }, []);
+
+  const initialExpandedPaths = useMemo(() => {
+    const rootId = getFileRootId();
+    const parts = rootId.split('/');
+    const paths: string[] = [];
+    for (let i = 1; i <= parts.length; i++) {
+      paths.push(parts.slice(0, i).join('/'));
+    }
+    return new Set(paths);
+  }, []);
+
+  const [expandedPaths, setExpandedPaths] = useState<Set<string>>(
+    initialExpandedPaths,
+  );
+
+  useEffect(() => {
+    setExpandedPaths((prev) => {
+      const newPaths = getParentPaths(currentPath);
+      const updated = new Set(prev);
+      newPaths.forEach((p) => updated.add(p));
+      return updated;
+    });
+  }, [currentPath, getParentPaths]);
+
+  const getDestinationFolderCopyHeader = useCallback(
+    (count: number, name: string | undefined) => {
+      return count === 1 && name
+        ? t('Copy "{{name}}" to', { name })
+        : t('Copy {{count}} items to', { count });
+    },
+    [t],
+  );
+
+  const getDestinationFolderMoveHeader = useCallback(
+    (count: number, name: string | undefined) => {
+      return count === 1 && name
+        ? t('Move "{{name}}" to', { name })
+        : t('Move {{count}} items to', { count });
+    },
+    [t],
+  );
+
+  const { bulkActionLabels, treeActionLabels, gridActionLabels } =
+    useFileManagerActionLabels(activeTab, t);
+
+  const renderDeleteConfirmationTitle = useCallback(
+    (files: string[]) => {
+      const count = files.length;
+      return count === 1
+        ? t('Confirm Deleting Item')
+        : t('Confirm Deleting Items');
+    },
+    [t],
+  );
+
+  const renderDeleteConfirmationContent = useCallback(
+    (files: string[]) => {
+      return (
+        <div className="px-6 py-3 text-sm">
+          <p className="mb-3 text-secondary">
+            {files.length === 1 ? (
+              <>
+                {t('Are you sure you want to delete')}{' '}
+                <span className="break-all text-primary">
+                  “{files[0].split('/').pop()}”?
+                </span>
+              </>
+            ) : (
+              <>
+                {t('Do you want to delete the following')}{' '}
+                <span className="text-primary">
+                  {files.length} {t('items?')}
+                </span>
+              </>
+            )}
+          </p>
+        </div>
+      );
+    },
+    [t],
+  );
+
+  const moveToFileHandler = useCallback(
+    (
+      movedItems: DialCopiedItem[],
+      sourceFolder: string,
+      destinationFolder: string,
+    ) => {
+      if (movedItems.length === 0) return;
+
+      movingFilesCountRef.current = movedItems.length;
+
+      dispatch(
+        FilesActions.moveFiles({
+          files: movedItems,
+          sourceFolder,
+          destinationFolder,
+        }),
+      );
+
+      const movedCurrent = movedItems.find(
+        (item) => item.sourceUrl === currentPath,
+      );
+
+      if (movedCurrent) {
+        setCurrentPath(undefined);
+      }
+    },
+    [dispatch, currentPath],
+  );
+
+  const handleSearchFiles = useCallback(
+    (folder: string) => {
+      dispatch(FilesActions.getFullListing({ folderPath: folder }));
+    },
+    [dispatch],
+  );
+
+  const renderOperationLoaderModal = useCallback(() => {
+    if (!isCopyingFiles && !isMovingFiles) {
+      return null;
+    }
+
+    const cancelHandler = isCopyingFiles
+      ? () => dispatch(FilesActions.cancelCopyingFiles())
+      : () => dispatch(FilesActions.cancelMovingFiles());
+
+    return {
+      title: t(isCopyingFiles ? 'Copying files' : 'Moving items'),
+      text: t('{{count}} items are being {{action}}…', {
+        count: movingFilesCountRef.current,
+        action: isCopyingFiles ? 'copied' : 'moved',
+      }),
+      onCancel: cancelHandler,
+    };
+  }, [isCopyingFiles, isMovingFiles, dispatch, t]);
+
+  const bulkActionsToolbarOptions = useMemo(
+    () => ({
+      actionLabels: bulkActionLabels,
+      getSelectionLabel(selectedCount: number) {
+        return selectedCount === 1
+          ? t('{{count}} item selected', { count: selectedCount })
+          : t('{{count}} items selected', { count: selectedCount });
+      },
+    }),
+    [bulkActionLabels, t],
+  );
+
+  const treeOptions = useMemo(
+    () => ({
+      expandedPaths,
+      title: t('Folder tree'),
+      collapsed: treeCollapsedState,
+      onCollapseChange: setTreeCollapsedState,
+      loadedPaths: loadedFoldersPaths,
+      actionLabels: treeActionLabels,
+    }),
+    [
+      expandedPaths,
+      t,
+      treeCollapsedState,
+      loadedFoldersPaths,
+      treeActionLabels,
+    ],
+  );
+
+  const fileMetadataPopupOptions = useMemo(
+    () => ({
+      title: t('Information'),
+      nameLabel: t('Name: '),
+      pathLabel: t('Path: '),
+      modifiedDateLabel: t('Modified: '),
+      sizeLabel: t('Size: '),
+      authorLabel: t('Author: '),
+      loading: isFileMetadataLoading,
+      fileMetadata: fileMetadata ?? undefined,
+    }),
+    [t, isFileMetadataLoading, fileMetadata],
+  );
+
+  const navigationPanelOptions = useMemo(
+    () => ({
+      searchable: true,
+    }),
+    [],
+  );
+
+  const dateOptions = useMemo(
+    () => ({
+      year: 'numeric' as const,
+      month: 'short' as const,
+      day: '2-digit' as const,
+    }),
+    [],
+  );
+
+  const gridOptions = useMemo(
+    () => ({
+      filterable: false,
+      dateLocale: 'en-US',
+      dateOptions: dateOptions,
+      actionLabels: gridActionLabels,
+      visibleColumns: visibleColumns,
+    }),
+    [dateOptions, gridActionLabels, visibleColumns],
+  );
+
+  const newActionLabels = useMemo(
+    () => ({
+      uploadFiles: t('Upload files'),
+      newFolder: t('New folder'),
+      uploadArchive: t('Upload archive'),
+    }),
+    [t],
+  );
+
+  const toolbarOptions = useMemo(
+    () => ({
+      tabs: tabs,
+      activeTab: activeTab,
+      onTabChange: handleTabChange,
+      newButtonVariant: ButtonVariant.Primary,
+      newActionLabels: newActionLabels,
+    }),
+    [tabs, activeTab, handleTabChange, newActionLabels],
+  );
+
+  const destinationFolderPopupOptions = useMemo(
+    () => ({
+      destinationFolderPath: destinationPath,
+      setDestinationFolderPath: setDestinationPath,
+      getCopyHeader: getDestinationFolderCopyHeader,
+      getMoveHeader: getDestinationFolderMoveHeader,
+    }),
+    [
+      destinationPath,
+      getDestinationFolderCopyHeader,
+      getDestinationFolderMoveHeader,
+    ],
+  );
+
+  const deleteConfirmationOptions = useMemo(
+    () => ({
+      titleRenderer: renderDeleteConfirmationTitle,
+      contentRenderer: renderDeleteConfirmationContent,
+    }),
+    [renderDeleteConfirmationTitle, renderDeleteConfirmationContent],
+  );
+
+  const handleCopyFiles = useCallback(
+    (copiedItems: DialCopiedItem[], destinationFolder: string) => {
+      if (copiedItems.length === 0) return;
+      movingFilesCountRef.current = copiedItems.length;
+      dispatch(
+        FilesActions.copyFiles({ files: copiedItems, destinationFolder }),
+      );
+    },
+    [dispatch],
+  );
+
+  const handleGetInfo = useCallback(
+    (item: { path: string } | null) => {
+      if (!item) return;
+      dispatch(FilesActions.getFileMetadata({ fileId: item.path }));
+    },
+    [dispatch],
+  );
+
+  const handleDeleteFiles = useCallback(
+    (deletedItems: DialDeletedItem[], folderUrl: string) => {
+      if (deletedItems.length === 0) return;
+      dispatch(
+        FilesActions.deleteFiles({
+          files: deletedItems,
+          folderUrl,
+        }),
+      );
+    },
+    [dispatch],
+  );
+
+  const handleDownloadFiles = useCallback(
+    (filesToDownload: DialFile[]) => {
+      if (filesToDownload.length === 0) return;
+      dispatch(
+        FilesActions.downloadFilesAsArchive({
+          files: filesToDownload,
+        }),
+      );
+    },
+    [dispatch],
+  );
+
+  const handleTableFileClick = useCallback((row: unknown) => {
+    void row;
+  }, []);
+
+  const handleUploadFiles = useCallback(
+    (filesToUpload: DialUploadFileItem[], destinationUrl: string) => {
+      if (filesToUpload.length === 0) return;
+      dispatch(
+        FilesActions.uploadFiles({
+          files: filesToUpload,
+          destinationUrl,
+        }),
+      );
+    },
+    [dispatch],
+  );
+
+  const handleCreateFolder = useCallback(
+    (file: DialUploadFileItem, folderPath: string) => {
+      dispatch(
+        FilesActions.uploadFiles({
+          files: [file],
+          destinationUrl: folderPath,
+        }),
+      );
+    },
+    [dispatch],
+  );
+
+  const handleUploadArchive = useCallback(
+    (archiveFile: File | null, name: string, destinationUrl: string) => {
+      if (!archiveFile) return;
+      dispatch(
+        FilesActions.uploadArchive({
+          archive: archiveFile,
+          name,
+          destinationUrl,
+        }),
+      );
+    },
+    [dispatch],
+  );
+
+  const handleUnshareFile = useCallback(
+    (file: { path: string; nodeType?: string }) => {
+      dispatch(
+        ShareActions.discardSharedWithMe({
+          resourceIds: [file.path],
+          featureType: FeatureType.File,
+          isFolder: file.nodeType === 'folder',
+        }),
+      );
+    },
+    [dispatch],
+  );
+
+  return {
+    currentPath,
+    setCurrentPath,
+    areFilesLoading,
+    areFoldersLoading,
+    isAnyOperationInProgress,
+    fileTreeItems,
+    rootFolder,
+    sharedByMePaths,
+    isLoadingSearchListing,
+    searchResultsUIKit,
+
+    operationLoaderModal: renderOperationLoaderModal(),
+
+    bulkActionsToolbarOptions,
+    treeOptions,
+    fileMetadataPopupOptions,
+    navigationPanelOptions,
+    gridOptions,
+    toolbarOptions,
+    destinationFolderPopupOptions,
+    deleteConfirmationOptions,
+
+    handleSearchFiles,
+    handleCopyFiles,
+    handleGetInfo,
+    moveToFileHandler,
+    handleDeleteFiles,
+    handleDownloadFiles,
+    handleTableFileClick,
+    handleUploadFiles,
+    handleCreateFolder,
+    handleUploadArchive,
+    handleUnshareFile,
+  };
+};
